refactor: modularize CSP handling and Ethereum provider logic
